requirements.txt
flask==2.3.2
mysql-connector-python==9.1.0